Debugging.Client: Do not mess with index when returning from GetFrame
This is a revert of 6d081b500dafc2443071c6223b5bf56cfd1cb06d.

The underlying issue is a desync between frames that end up in the
client cache and what was actually returned by the runtime. Such issues
should be exposed and fixed instead of returning bad data up the stack.